Allow optionally building with no '.git' dir
This is primarily motivated by packaging maptool with nix. In nix, it's
encouraged to avoid leaving around the '.git' for git repositories
because it can lead to some subtle nondeterministic behavior (i.e. it
doesn't always hash to the same thing, even though you're checked out to
the same commit).

Further, this allows building the source archives that are on the
'releases' page, though it admittedly does take a rather clumsy
invocation of gradle.

Other linux distros likely have similar desires to be able to compile
the program from a source tarball.